refactor(suite-native): extract navigate-to-staking-detail-with-analytics hook
